Add security to viewst to allow only superusers to change store contents

+++ products/views.py
@@ -5,6 +5,7 @@ from django.db.models import Q
 from django.contrib import messages
 from django.db.models.functions import Lower
 from .models import Product, Category
+from django.contrib.auth.decorators import login_required
 
 def all_products(request):
     """A view to return all products"""
@@ -81,9 +82,14 @@ def product_details(request, product_id):
 
     return render(request, 'products/product_details.html', context)
 
-
+@login_required
 def add_product(request):
     """Add a product to the database"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, it looks like you are not logged in as a super user. \
+                                This action is not allowed.')
+        return redirect(reverse('home'))
+
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -104,9 +110,16 @@ def add_product(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_product(request, product_id):
     """Edit a product in the shop"""
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, it looks like you are not logged in as a super user. \
+                                This action is not allowed.')
+        return redirect(reverse('home'))
+
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -128,9 +141,16 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
     """Delete a product in the shop"""
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, it looks like you are not logged in as a super user. \
+                                This action is not allowed.')
+        return redirect(reverse('home'))
+
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted successfully!')
